Remove 'magic' `swift` and `scality-sproxyd-client` installation
To reduce the number of accidental installations of whatever Swift
release we point to in `requirements.txt`, this commit removes it
altogether, whilst keeping the dependencies.

When running `pip -r requirements.txt` on a system where Swift or
scality-sproxyd-client are not installed, this will result in an error
and require manual installation.