Hash the account, container and object name
With this, no need to change the Apache config flag 'AllowEncodedSlahes'
and we also get consistent behavior with all web servers.

Warning: this is completely backward incompatible. No Swift object put before
this commit will be reachable once this patch is merged. Users of the 0.3
branch of this project should do a full copy of their Swift cluster if they
want to migrate to this branch.