Change USER from stack (doesn't exist) to swift
This solves an issue with the container image failing the certification scan. The rpm_list_successful runs the `rpm -qa` command in docker on the running container, however this container fails to run in the current configuration, since the user "stack" does not exist in the container image.

contributed by jsm84